+++ TidalPy/Utilities_x/logging_x/__init__.py
@@ -1,0 +1,5 @@
+"""TidalPy C++ logging interface (spdlog wrapper)."""
+
+from TidalPy.Utilities_x.logging_x.logger import init_logger, set_log_level, shutdown_logger
+
+__all__ = ["init_logger", "set_log_level", "shutdown_logger"]